refactor: (19.08.23) Moved creation of balance change to a separate class (#80)
